requirements.txt
discord==2.3.2
google-genai==1.74.0
python-dotenv==1.2.2